modules: introduce MODULE_xxx_UPLOAD_REMOTE_SUPPORT variable
This new variable need to be declared in module file.
Remote upload is current not implemented by any module (the only one example was megaupload, but it's gone now).

In upload module function, you can easily detect if $FILE is a local file or a remote one:

<code>
if match_remote_url "$FILE"; then
    ...
    ... treat remote file
    ...
</code>