extabit (upload): alternate, more robust parsing of download URL
After upload Extabit redirects to the file's download page. The
old method used to parse the <form> on that page for the download
link. But this form is not available when the user is not allowed
to download the file - file >400MB require a premium account.

The new method parses the download link from another part of the
page that should always be there.